base/requirements.txt
Pillow
numpy
requests
click